Pruebas unitarias que se han realizado hasta los momentos. Tambien se incluyeron las clases con algunas modificaciones.

+++ Clases/Cliente.java
@@ -33,7 +33,7 @@ public class Cliente {
      * Permite agregar un cliente nuevo a la base de datos.
      * @throws SQLException puede lanzar una excepcion de sql
      */
-    void registrarCliente() throws SQLException {
+    void registrarCliente() {
         // Conectamos con la base de datos
         try (Connection conn = Conexion.obtenerConn()) {
             
@@ -43,7 +43,7 @@ public class Cliente {
             // Se agrega el nuevo cliente a la tabla cliente y se agregan sus
             // telefonos a la tabla telefono.
             
-            st.execute("insert into cliente values ('" + this.cedula.toString() + "', '"
+            st.execute("insert into cliente values ('" + this.cedula + "', '"
                     + this.nombre + "', '" + this.direccion + "');");
 
             for (int i=0; i < telefonos.toArray().length; ++i) {
@@ -79,7 +79,7 @@ public class Cliente {
             
             // Buscamos los telefonos del cliente en la base de datos
             ResultSet rs = st.executeQuery("select numero from telefono where "
-                    + "ci = '" + cedula.toString() + "';");
+                    + "ci = " + cedula.toString() + ";");
 
             // Si los conseguimos es porque existe el cliente
             if (rs.next()) {
@@ -101,8 +101,9 @@ public class Cliente {
                 rs.close();
                 
                 // Buscamos al cliente en la base de datos
-                rs = st.executeQuery("select ci, nombre, direccion from cliente where ci = '" + 
-                    cedula.toString() + "';");
+                
+                rs = st.executeQuery("select ci, nombre, direccion from cliente where ci = " + 
+                    cedula.toString() + ";");
             
                 rs.next();
               
@@ -170,17 +171,20 @@ public class Cliente {
         try (Connection conn = Conexion.obtenerConn()) {
             Statement st;
             st = conn.createStatement();
-                        
+            
+            
+            
             // Eliminamos los telefonos del cliente
-            st.execute("delete from Telefono where ci ='"+this.cedula+"';");
+            st.execute("delete from Telefono where ci ="+this.cedula+";");
 
             // Eliminamos los productos que un clinte posee
-            for (int i=0; i < productos.toArray().length; ++i) {
-                ((Producto) productos.get(i)).eliminarProducto();
-            }            
+            if (productos != null) {
+                for (int i=0; i < productos.toArray().length; ++i) {
+                    ((Producto) productos.get(i)).eliminarProducto();
+                }        
+            }
             
-            // Eliminamos al cliente
-            st.execute("delete from Cliente where ci ='"+this.cedula+"';");
+            st.execute("delete from Cliente where ci ="+this.cedula+";");
             
           } catch (SQLException ex) {
               // Si hay una excepcion se imprime un mensaje
